Add verification modal for event deletion
