+++ met-web/src/components/engagement/form/EngagementWidgets/Events/VirtualEventInfoPaper.tsx
@@ -8,26 +8,10 @@ import { When } from 'react-if';
 import { formatDate } from 'components/common/dateHelper';
 import { EventInfoPaperProps } from './EventInfoPaper';
 import { EventsContext } from './EventsContext';
-import { deleteEvent } from 'services/widgetService/EventService';
-import { useAppDispatch } from 'hooks';
-import { openNotification } from 'services/notificationService/notificationSlice';
 
 const VirtualEventInfoPaper = ({ event, removeEvent, ...rest }: EventInfoPaperProps) => {
     const eventItem = event.event_items[0];
-    const dispatch = useAppDispatch();
-    const { handleChangeEventToEdit, handleEventDrawerOpen, widget } = useContext(EventsContext);
-
-    const handleRemoveEvent = async (event_id: number) => {
-        try {
-            if (widget) {
-                await deleteEvent(widget.id, event_id);
-                removeEvent(event_id);
-                dispatch(openNotification({ severity: 'success', text: 'The event was removed successfully' }));
-            }
-        } catch (error) {
-            dispatch(openNotification({ severity: 'error', text: 'An error occurred while trying to remove event' }));
-        }
-    };
+    const { handleChangeEventToEdit, handleEventDrawerOpen } = useContext(EventsContext);
 
     return (
         <MetWidgetPaper elevation={1} {...rest}>
@@ -103,7 +87,7 @@ const VirtualEventInfoPaper = ({ event, removeEvent, ...rest }: EventInfoPaperPr
                     </Grid>
                     <Grid item xs={6}>
                         <IconButton
-                            onClick={() => handleRemoveEvent(event.id)}
+                            onClick={() => removeEvent(event.id)}
                             sx={{ padding: 1, margin: 0 }}
                             color="inherit"
                             aria-label="delete-icon"
